[REFACTOR] Move generic function to tests/function.py.
