+++ tests/create_tests.py
@@ -6,32 +6,6 @@ from git import Repo
 import requests
 
 
-def project_name2env(project_name: str) -> str:
-    """Convert the project name to a valid ENV var name.
-
-    The ENV name for the project must match the regex `[a-zA-Z_]+[a-zA-Z0-9_]*`.
-
-    Parameters
-    ----------
-    project_name: str
-        Name of the project.
-
-    Return
-    ------
-    project_env: str
-        A valid ENV name for the project.
-    """
-    project_name = project_name.replace("-", "_")
-    project_env = re.sub("[_]+", "_", project_name)  # Remove consecutive `_`
-    project_env = re.sub("[^a-zA-Z0-9_]", "", project_env)
-
-    # Env var cannot start with number
-    if re.compile("[0-9]").match(project_env[0]):
-        project_env = "_" + project_env
-
-    return project_env.upper()
-
-
 def minimal_tests(datasets: List[str], pr_files: List[str]):
     """Return the minimal test set for a pull request.
